+++ build-tools/utils/aab.d.ts
@@ -14,7 +14,6 @@ export function assembleAab(options: {
     contentDir: string;
     app: ApkAppIdentity;
     key: SigningKey;
-    abi: string;
 }): Buffer;
 
 export function aabFileName(appId: string): string;